# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Collaborative coding platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Projects ─────────────────────────────────────────────────────
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for current user
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, owned, shared, recent]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/members:
    get:
      operationId: listProjectMembers
      tags: [projects]
      summary: List project members
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
  /projects/{projectId}/invite:
    post:
      operationId: inviteMember
      tags: [projects]
      summary: Invite a member to project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "200":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
  /projects/{projectId}/members/{userId}:
    patch:
      operationId: updateMemberRole
      tags: [projects]
      summary: Update member role
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"

    delete:
      operationId: removeMember
      tags: [projects]
      summary: Remove member from project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /projects/{projectId}/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get project analytics stats
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

  # ── Files ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string

    # ── Projects ──────────────────────────────────────────────────
    Project:
      type: object
      required: [id, name, ownerId, language, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: string
        ownerId:
          type: integer
        ownerName:
          type: ["string", "null"]
        ownerAvatarUrl:
          type: ["string", "null"]
        isPublic:
          type: boolean
        memberCount:
          type: integer
        fileCount:
          type: integer
        role:
          type: ["string", "null"]
        lastOpenedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [name, language]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        language:
          type: string
        isPublic:
          type: boolean
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        language:
          type: string
        isPublic:
          type: boolean
    ProjectMember:
      type: object
      required: [userId, projectId, role, joinedAt]
      properties:
        userId:
          type: integer
        projectId:
          type: integer
        role:
          type: string
          enum: [owner, admin, editor, viewer]
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        joinedAt:
          type: string
          format: date-time
    InviteInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, editor, viewer]
    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, editor, viewer]
    ProjectStats:
      type: object
      required: [totalEdits, totalCommits, activeUsers, filesModifiedToday, chatActivity, weeklyActivity]
      properties:
        totalEdits:
          type: integer
        totalCommits:
          type: integer
        activeUsers:
          type: integer
        filesModifiedToday:
          type: integer
        chatActivity:
          type: integer
        mostActiveUser:
          type: ["string", "null"]
        mostActiveUserAvatarUrl:
          type: ["string", "null"]
        weeklyActivity:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyActivityPoint"
    WeeklyActivityPoint:
      type: object
      required: [day, edits, messages]
      properties:
        day:
          type: string
        edits:
          type: integer
        messages:
          type: integer

    # ── Files ─────────────────────────────────────────────────────
...