# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Collaborative coding platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WeeklyActivityPoint } from './weeklyActivityPoint';

export interface ProjectStats {
  totalEdits: number;
  totalCommits: number;
  activeUsers: number;
  filesModifiedToday: number;
  chatActivity: number;
  /** @nullable */
  mostActiveUser?: string | null;
  /** @nullable */
  mostActiveUserAvatarUrl?: string | null;
  weeklyActivity: WeeklyActivityPoint[];
}